refactor: 예외 처리 한 곳에서 하도록 수정, 500 예외 메시지 수정
