+++ src/main/java/roomescape/presentation/exception/GlobalExceptionHandler.java
@@ -1,8 +1,5 @@
 package roomescape.presentation.exception;
 
-import java.util.NoSuchElementException;
-import java.util.stream.Collectors;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
@@ -12,13 +9,17 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.application.exception.AuthException;
+import roomescape.domain.exception.PastReservationException;
 import roomescape.presentation.dto.response.ErrorResponse;
+
+import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
     private static final Logger log = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-    private static final String PREFIX_FOR_ERROR_MESSAGE = "알 수 없는 문제 발생, ";
+    private static final String INTERNAL_SERVER_ERROR_MESSAGE = "알 수 없는 문제 발생했습니다.";
 
     @ExceptionHandler(NoSuchElementException.class)
     public ResponseEntity<ErrorResponse> handleNoSuchElementException(NoSuchElementException e) {
@@ -61,9 +62,16 @@ public class GlobalExceptionHandler {
         return ResponseEntity.badRequest().body(errorResponse);
     }
 
+    @ExceptionHandler(PastReservationException.class)
+    public ResponseEntity<ErrorResponse> handlePastReservationException(PastReservationException e) {
+        String message = e.getMessage();
+        ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.BAD_REQUEST, message);
+        return ResponseEntity.badRequest().body(errorResponse);
+    }
+
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ErrorResponse> handleAllExceptions(Exception e) {
-        String message = PREFIX_FOR_ERROR_MESSAGE + e.getMessage();
+        String message = INTERNAL_SERVER_ERROR_MESSAGE;
         log.error(message, e);
         ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.INTERNAL_SERVER_ERROR, message);
         return ResponseEntity.internalServerError().body(errorResponse);
